new(dev): Add basic service worker config and build
This should enable the site to be a decent PWA.

It only precaches the images in `assets/img` (logos and icons) to
prevent a huge bloated install/download.
This means that images will be broken in posts.

Otherwise, it precaches (makes offline) all html pages and fonts and css
and js and json.
This should allow someone to read and search and such.

Ref: https://fredrickb.com/2019/07/25/turning-jekyll-site-into-a-progressive-web-app/